requirements.txt
absl-py==0.9.0
apex==0.1
astor==0.8.1
astunparse==1.6.3
attrs==19.3.0
backcall==0.1.0
beautifulsoup4==4.8.2
bleach==3.1.3
blis==0.4.1
cachetools==4.1.0
catalogue==1.0.0
certifi==2020.6.20
chardet==3.0.4
click==7.1.2
comet-git-pure==0.19.16
comet-ml==3.1.7
configobj==5.0.6
cycler==0.10.0
cymem==2.0.3
decorator==4.4.2
defusedxml==0.6.0
dill==0.3.1.1
en-core-web-sm==2.2.5
entrypoints==0.3
everett==1.0.2
fast-bert==1.7.1
fastprogress==0.2.3
filelock==3.0.12
future==0.18.2
gast==0.2.2
google-auth==1.14.3
google-auth-oauthlib==0.4.1
google-pasta==0.2.0
googleapis-common-protos==1.51.0
grpcio==1.28.1
h5py==2.10.0
idna==2.9
importlib-metadata==1.7.0
ipykernel==5.1.4
ipython==7.13.0
ipython-genutils==0.2.0
ipywidgets==7.5.1
jedi==0.17.0
Jinja2==2.11.1
joblib==0.14.1
jsonschema==3.0.2
jupyter==1.0.0
jupyter-client==6.0.0
jupyter-console==6.1.0
jupyter-core==4.6.3
Keras==2.3.1
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.1
kiwisolver==1.1.0
Markdown==3.2.2
MarkupSafe==1.1.1
matplotlib==3.2.0
mistune==0.8.4
mkl-fft==1.0.15
mkl-random==1.1.0
mkl-service==2.3.0
more-itertools==8.4.0
murmurhash==1.0.2
nbconvert==5.6.1
nbformat==5.0.4
netifaces==0.10.9
nltk==3.4.5
notebook==6.0.3
numpy==1.18.1
nvidia-ml-py3==7.352.0
oauthlib==3.1.0
opt-einsum==3.2.1
packaging==20.4
pactools==0.2.0b0
pandas==1.0.1
pandocfilters==1.4.2
parso==0.7.0
pexpect==4.8.0
pickleshare==0.7.5
Pillow==7.1.2
plac==1.1.3
pluggy==0.13.1
portalocker==1.7.0
preshed==3.0.2
prometheus-client==0.7.1
promise==2.3
prompt-toolkit==3.0.4
protobuf==3.11.3
psutil==5.7.0
ptyprocess==0.6.0
py==1.9.0
pyasn1==0.4.8
pyasn1-modules==0.2.8
Pygments==2.6.1
pyparsing==2.4.7
pyrsistent==0.15.7
pytest==5.4.3
python-box==4.2.3
python-crfsuite==0.9.7
python-dateutil==2.8.1
pytorch-lamb==1.0.0
pytorch-lightning==0.7.3
pytrec-eval==0.4
pytz==2020.1
PyYAML==5.3.1
pyzmq==19.0.0
qtconsole==4.7.1
QtPy==1.9.0
regex==2020.5.7
requests==2.23.0
requests-oauthlib==1.3.0
rouge-score==0.0.3
rsa==4.0
ruamel.yaml==0.16.10
ruamel.yaml.clib==0.2.0
sacrebleu==1.4.9
sacremoses==0.0.43
scikit-learn==0.22.2.post1
scipy==1.4.1
Send2Trash==1.5.0
sentencepiece==0.1.86
seqeval==0.0.12
six==1.14.0
sklearn==0.0
sklearn-crfsuite==0.3.6
soupsieve==2.0
spacy==2.2.4
srsly==1.0.2
stanza==1.0.1
tabulate==0.8.7
tensorboard==2.0.2
tensorboard-plugin-wit==1.6.0.post3
tensorboardX==2.0
tensorflow==2.0.0
tensorflow-datasets==3.1.0
tensorflow-estimator==2.0.1
tensorflow-gpu==2.0.0
tensorflow-metadata==0.22.0
termcolor==1.1.0
terminado==0.8.3
testpath==0.4.4
thinc==7.4.0
tokenizers==0.7.0
toml==0.10.1
torch==1.5.0
torchvision==0.4.0
tornado==6.0.4
tqdm==4.42.1
traitlets==4.3.3
transformers==2.11.0
typing==3.7.4.1
typing-extensions==3.7.4.2
urllib3==1.25.9
wasabi==0.6.0
wcwidth==0.1.9
webencodings==0.5.1
websocket-client==0.57.0
Werkzeug==1.0.1
widgetsnbextension==3.5.1
wrapt==1.12.1
wurlitzer==2.0.0
zipp==3.1.0
